Fix test_registration and Mesh type. Update README.
Fix Mesh type to have edges attribute as a sparse matrix of Int64s.
Fix test_registration to work with that type update.

Update README with example on how to use meshwarp.

+++ src/mesh.jl
@@ -23,7 +23,7 @@
 type Mesh
   src_nodes::Array 
   dst_nodes::Array
-  edges::SparseMatrixCSC{Float64, Int64}
+  edges::SparseMatrixCSC{Int64, Int64}
 end
 
 Mesh() = Mesh([], [], spzeros(0,0))
